Method to read, subset and sample expanded data

This method is used on trial_sequence objects to read, subset and sample expanded data.

sample_controls(
  object,
  p_control,
  period = NULL,
  subset_condition = NULL,
  seed = NULL
)

# S4 method for trial_sequence
sample_controls(
  object,
  p_control,
  period = NULL,
  subset_condition = NULL,
  seed = NULL
)

Arguments

object

An object of class trial_sequence.

p_control

Probability of selecting a control, default is 0.01.

period

An integerish vector of non-zero length to select trial period(s) or NULL (default) to select all trial periods.

subset_condition

A string or NULL (default). subset_condition will be translated to a call (in case the expanded data is saved as a data.table or in the csv format) or to a SQL-query (in case the expanded data is saved as a duckdb file).

The operators "==", "!=", ">", ">=", "<", "<=", %in%", "&", "|" are supported. Numeric vectors can be written as c(1, 2, 3) or 1:3. Variables are not supported.

Note: Make sure numeric vectors written as 1:3 are surrounded by spaces, e.g. a %in% c( 1:4 , 6:9 ), otherwise the code will fail.

seed

An integer seed or NULL (default).

Note: The same seed will return a different result depending on the class of the te_datastore object contained in the trial_sequence object.

Value

A data.frame of class data.table.

Examples

# create a trial_sequence-class object
trial_itt_dir <- file.path(tempdir(), "trial_itt")
dir.create(trial_itt_dir)
trial_itt <- trial_sequence(estimand = "ITT") |>
  set_data(data = data_censored) |>
  set_outcome_model(adjustment_terms = ~ x1 + x2)

trial_itt_csv <- set_expansion_options(
  trial_itt,
  output = save_to_csv(file.path(trial_itt_dir, "trial_csvs")),
  chunk_size = 500
) |>
  expand_trials()

# sample_controls default behaviour returns all trial_periods
sample_controls(trial_itt_csv, p_control = 0.01)
#>        id trial_period followup_time outcome weight treatment    x1         x2
#>     <int>        <int>         <int>   <int>  <int>     <int> <int>      <num>
#>  1:    15            0             0       1      1         1     0 -0.7365256
#>  2:    32            0             0       1      1         1     1  1.9861380
#>  3:    55            0             1       0     NA         1     1  0.4480097
#>  4:    61            0             2       1     NA         0     1 -0.6123324
#>  5:    74            2             3       0     NA         0     0  0.1175068
#>  6:    54            2             3       0     NA         0     0 -0.3877549
#>  7:    68            0             4       1     NA         0     0 -0.3919838
#>  8:    96            0             5       0     NA         1     0 -1.9542363
#>  9:    49            1             6       1     NA         0     0 -2.2687065
#> 10:    74            2             6       0     NA         1     0  0.1175068
#> 11:    29            0             7       1     NA         1     0  0.4037240
#> 12:    49            0             7       1     NA         0     0  1.1572111
#> 13:    89            0             7       1     NA         0     1  0.4128518
#> 14:    99            0             7       1     NA         0     1 -0.3463778
#> 15:    38            1             7       1     NA         0     1 -1.3912346
#> 16:    74            3             7       0     NA         0     0 -0.8480194
#> 17:    38            0             8       1     NA         0     0 -0.2469111
#> 18:    34            1             8       0     NA         1     1  0.2416765
#> 19:    74            6             9       0     NA         0     1 -0.7825051
#> 20:    54            8            10       0     NA         0     1  0.1536732
#> 21:    59            1            12       0     NA         0     0 -1.1828232
#> 22:    51            0            13       0     NA         0     1 -0.1352684
#> 23:    71            1            13       1     NA         0     0 -0.3970024
#> 24:    71            0            14       1     NA         0     0 -0.2678913
#> 25:    98            0            14       1     NA         0     1  1.3923395
#> 26:    95            3            14       0     NA         1     1  0.6871256
#> 27:    74            3            14       0     NA         0     0 -0.8480194
#> 28:    53            0            15       0     NA         0     1 -1.1822289
#>        id trial_period followup_time outcome weight treatment    x1         x2
#>     assigned_treatment sample_weight
#>                  <int>         <num>
#>  1:                  1             1
#>  2:                  1             1
#>  3:                  1           100
#>  4:                  1             1
#>  5:                  0           100
#>  6:                  0           100
#>  7:                  1             1
#>  8:                  0           100
#>  9:                  1             1
#> 10:                  0           100
#> 11:                  1             1
#> 12:                  0             1
#> 13:                  1             1
#> 14:                  1             1
#> 15:                  1             1
#> 16:                  0           100
#> 17:                  0             1
#> 18:                  1           100
#> 19:                  0           100
#> 20:                  0           100
#> 21:                  0           100
#> 22:                  1           100
#> 23:                  1             1
#> 24:                  0             1
#> 25:                  1             1
#> 26:                  1           100
#> 27:                  0           100
#> 28:                  0           100
#>     assigned_treatment sample_weight

# sample_controls can subset the data before sampling
sample_controls(
  trial_itt_csv,
  p_control = 0.2,
  period = 1:20,
  subset_condition = "followup_time %in% 1:20 & treatment == 1",
)
#>        id trial_period followup_time outcome weight treatment    x1          x2
#>     <int>        <int>         <int>   <int> <lgcl>     <int> <int>       <num>
#>  1:    21            3             1       0     NA         1     0  1.87509512
#>  2:    70            1             2       0     NA         1     0 -0.52391800
#>  3:    27            1             2       0     NA         1     0  0.68971049
#>  4:    83            1             2       0     NA         1     0 -0.11345112
#>  5:    49            1             2       0     NA         1     0 -2.26870654
#>  6:    60            2             2       0     NA         1     1  0.79983377
#>  7:    74            7             2       0     NA         1     0 -0.58362639
#>  8:    34            1             3       0     NA         1     1  0.24167648
#>  9:    53            1             3       0     NA         1     1 -0.18557643
#> 10:    16            1             4       0     NA         1     1 -0.69264361
#> 11:    60            2             4       0     NA         1     1  0.79983377
#> 12:    95            2             4       0     NA         1     1  0.01992213
#> 13:    74            4             4       0     NA         1     1 -0.97471771
#> 14:    71            1             5       0     NA         1     0 -0.39700243
#> 15:    44            3             5       0     NA         1     1 -0.59779120
#> 16:    33            1             6       0     NA         1     0 -0.55141438
#> 17:    53            1             6       0     NA         1     1 -0.18557643
#> 18:    20            1             6       0     NA         1     0 -0.42516753
#> 19:    34            1             6       0     NA         1     1  0.24167648
#> 20:    44            2             6       0     NA         1     0  0.84818522
#> 21:    95            3             7       0     NA         1     1  0.68712558
#> 22:    95            1             8       0     NA         1     0 -1.52782784
#> 23:    59            2             8       0     NA         1     1  0.68201510
#> 24:    53            1             9       0     NA         1     1 -0.18557643
#> 25:    54            8             9       0     NA         1     1  0.15367316
#> 26:    96            1            10       0     NA         1     0 -1.08532471
#> 27:    33            1            10       0     NA         1     0 -0.55141438
#> 28:    44            3            10       0     NA         1     1 -0.59779120
#> 29:    95            3            10       0     NA         1     1  0.68712558
#> 30:    95            1            11       0     NA         1     0 -1.52782784
#> 31:    21            3            11       0     NA         1     0  1.87509512
#> 32:    21            2            12       0     NA         1     0 -0.80501051
#> 33:    44            3            12       0     NA         1     1 -0.59779120
#> 34:    59            1            13       0     NA         1     0 -1.18282317
#> 35:    95            2            13       0     NA         1     1  0.01992213
#> 36:    59            2            13       0     NA         1     1  0.68201510
#> 37:    44            2            13       0     NA         1     0  0.84818522
#> 38:    14            1            14       0     NA         1     0  0.51664948
#> 39:    44            4            14       0     NA         1     0 -0.41225502
#> 40:    59            2            15       0     NA         1     1  0.68201510
#> 41:    95            2            15       0     NA         1     1  0.01992213
#> 42:    44            3            15       0     NA         1     1 -0.59779120
#> 43:    44            4            15       0     NA         1     0 -0.41225502
#> 44:    54            1            16       0     NA         1     0 -1.43263029
#> 45:    59            1            16       0     NA         1     0 -1.18282317
#>        id trial_period followup_time outcome weight treatment    x1          x2
#>     assigned_treatment sample_weight
#>                  <int>         <num>
#>  1:                  1             5
#>  2:                  1             5
#>  3:                  0             5
#>  4:                  0             5
#>  5:                  1             5
#>  6:                  1             5
#>  7:                  0             5
#>  8:                  1             5
#>  9:                  1             5
#> 10:                  1             5
#> 11:                  1             5
#> 12:                  0             5
#> 13:                  0             5
#> 14:                  1             5
#> 15:                  0             5
#> 16:                  1             5
#> 17:                  1             5
#> 18:                  0             5
#> 19:                  1             5
#> 20:                  0             5
#> 21:                  1             5
#> 22:                  0             5
#> 23:                  1             5
#> 24:                  1             5
#> 25:                  0             5
#> 26:                  1             5
#> 27:                  1             5
#> 28:                  0             5
#> 29:                  1             5
#> 30:                  0             5
#> 31:                  1             5
#> 32:                  0             5
#> 33:                  0             5
#> 34:                  0             5
#> 35:                  0             5
#> 36:                  1             5
#> 37:                  0             5
#> 38:                  1             5
#> 39:                  1             5
#> 40:                  1             5
#> 41:                  0             5
#> 42:                  0             5
#> 43:                  1             5
#> 44:                  0             5
#> 45:                  0             5
#>     assigned_treatment sample_weight

# delete after use
unlink(trial_itt_dir, recursive = TRUE)