Method to save expanded data

This method is used internally by expand_trials to save the data to the "datastore" defined in set_expansion_options.

save_expanded_data(object, data)

# S4 method for te_datastore_csv
save_expanded_data(object, data)

# S4 method for te_datastore_datatable
save_expanded_data(object, data)

# S4 method for te_datastore_duckdb
save_expanded_data(object, data)

Arguments

object

An object of class te_datastore or a child class.

data

A data frame containing the expanded trial data. The columns trial_period and id are present, which may be used in methods to save the data in an optimal way, such as with indexes, keys or separate files.

Value

An updated object with the data stored. Notably object@N should be increased

Examples

temp_dir <- tempfile("csv_dir_")
dir.create(temp_dir)
datastore <- save_to_csv(temp_dir)
data(vignette_switch_data)
save_expanded_data(datastore, vignette_switch_data)
#> An object of class "te_datastore_csv"
#> Slot "path":
#> [1] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157"
#> 
#> Slot "files":
#>   [1] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_261.csv"
#>   [2] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_262.csv"
#>   [3] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_263.csv"
#>   [4] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_264.csv"
#>   [5] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_265.csv"
#>   [6] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_266.csv"
#>   [7] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_267.csv"
#>   [8] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_268.csv"
#>   [9] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_269.csv"
#>  [10] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_270.csv"
#>  [11] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_271.csv"
#>  [12] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_272.csv"
#>  [13] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_273.csv"
#>  [14] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_274.csv"
#>  [15] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_275.csv"
#>  [16] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_276.csv"
#>  [17] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_277.csv"
#>  [18] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_278.csv"
#>  [19] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_279.csv"
#>  [20] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_280.csv"
#>  [21] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_281.csv"
#>  [22] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_282.csv"
#>  [23] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_283.csv"
#>  [24] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_284.csv"
#>  [25] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_285.csv"
#>  [26] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_286.csv"
#>  [27] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_287.csv"
#>  [28] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_288.csv"
#>  [29] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_289.csv"
#>  [30] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_290.csv"
#>  [31] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_291.csv"
#>  [32] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_292.csv"
#>  [33] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_293.csv"
#>  [34] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_294.csv"
#>  [35] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_295.csv"
#>  [36] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_296.csv"
#>  [37] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_297.csv"
#>  [38] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_298.csv"
#>  [39] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_299.csv"
#>  [40] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_300.csv"
#>  [41] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_301.csv"
#>  [42] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_302.csv"
#>  [43] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_303.csv"
#>  [44] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_304.csv"
#>  [45] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_305.csv"
#>  [46] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_306.csv"
#>  [47] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_307.csv"
#>  [48] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_308.csv"
#>  [49] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_309.csv"
#>  [50] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_310.csv"
#>  [51] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_311.csv"
#>  [52] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_312.csv"
#>  [53] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_313.csv"
#>  [54] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_314.csv"
#>  [55] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_315.csv"
#>  [56] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_316.csv"
#>  [57] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_317.csv"
#>  [58] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_318.csv"
#>  [59] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_319.csv"
#>  [60] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_320.csv"
#>  [61] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_321.csv"
#>  [62] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_322.csv"
#>  [63] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_323.csv"
#>  [64] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_324.csv"
#>  [65] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_325.csv"
#>  [66] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_326.csv"
#>  [67] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_327.csv"
#>  [68] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_328.csv"
#>  [69] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_329.csv"
#>  [70] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_330.csv"
#>  [71] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_331.csv"
#>  [72] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_332.csv"
#>  [73] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_333.csv"
#>  [74] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_334.csv"
#>  [75] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_335.csv"
#>  [76] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_336.csv"
#>  [77] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_337.csv"
#>  [78] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_338.csv"
#>  [79] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_339.csv"
#>  [80] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_340.csv"
#>  [81] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_341.csv"
#>  [82] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_342.csv"
#>  [83] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_343.csv"
#>  [84] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_344.csv"
#>  [85] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_356.csv"
#>  [86] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_357.csv"
#>  [87] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_358.csv"
#>  [88] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_359.csv"
#>  [89] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_345.csv"
#>  [90] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_346.csv"
#>  [91] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_347.csv"
#>  [92] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_348.csv"
#>  [93] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_349.csv"
#>  [94] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_350.csv"
#>  [95] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_351.csv"
#>  [96] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_352.csv"
#>  [97] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_353.csv"
#>  [98] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_354.csv"
#>  [99] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_355.csv"
#> [100] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_360.csv"
#> [101] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_361.csv"
#> [102] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_221.csv"
#> [103] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_222.csv"
#> [104] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_223.csv"
#> [105] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_224.csv"
#> [106] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_225.csv"
#> [107] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_226.csv"
#> [108] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_227.csv"
#> [109] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_228.csv"
#> [110] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_229.csv"
#> [111] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_230.csv"
#> [112] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_231.csv"
#> [113] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_232.csv"
#> [114] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_233.csv"
#> [115] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_234.csv"
#> [116] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_235.csv"
#> [117] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_236.csv"
#> [118] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_237.csv"
#> [119] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_238.csv"
#> [120] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_239.csv"
#> [121] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_240.csv"
#> [122] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_241.csv"
#> [123] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_242.csv"
#> [124] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_243.csv"
#> [125] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_244.csv"
#> [126] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_245.csv"
#> [127] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_246.csv"
#> [128] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_247.csv"
#> [129] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_248.csv"
#> [130] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_249.csv"
#> [131] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_250.csv"
#> [132] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_251.csv"
#> [133] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_252.csv"
#> [134] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_253.csv"
#> [135] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_254.csv"
#> [136] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_255.csv"
#> [137] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_256.csv"
#> [138] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_257.csv"
#> [139] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_258.csv"
#> [140] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_259.csv"
#> [141] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_260.csv"
#> [142] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_115.csv"
#> [143] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_116.csv"
#> [144] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_117.csv"
#> [145] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_118.csv"
#> [146] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_119.csv"
#> [147] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_120.csv"
#> [148] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_121.csv"
#> [149] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_122.csv"
#> [150] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_123.csv"
#> [151] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_124.csv"
#> [152] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_125.csv"
#> [153] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_126.csv"
#> [154] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_127.csv"
#> [155] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_128.csv"
#> [156] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_129.csv"
#> [157] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_130.csv"
#> [158] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_131.csv"
#> [159] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_132.csv"
#> [160] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_133.csv"
#> [161] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_134.csv"
#> [162] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_135.csv"
#> [163] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_136.csv"
#> [164] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_137.csv"
#> [165] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_138.csv"
#> [166] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_139.csv"
#> [167] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_140.csv"
#> [168] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_141.csv"
#> [169] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_142.csv"
#> [170] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_143.csv"
#> [171] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_144.csv"
#> [172] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_145.csv"
#> [173] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_146.csv"
#> [174] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_147.csv"
#> [175] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_148.csv"
#> [176] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_149.csv"
#> [177] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_150.csv"
#> [178] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_151.csv"
#> [179] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_152.csv"
#> [180] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_153.csv"
#> [181] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_154.csv"
#> [182] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_155.csv"
#> [183] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_156.csv"
#> [184] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_157.csv"
#> [185] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_158.csv"
#> [186] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_159.csv"
#> [187] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_160.csv"
#> [188] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_161.csv"
#> [189] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_162.csv"
#> [190] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_163.csv"
#> [191] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_164.csv"
#> [192] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_165.csv"
#> [193] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_166.csv"
#> [194] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_167.csv"
#> [195] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_168.csv"
#> [196] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_169.csv"
#> [197] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_170.csv"
#> [198] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_171.csv"
#> [199] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_172.csv"
#> [200] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_173.csv"
#> [201] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_174.csv"
#> [202] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_175.csv"
#> [203] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_176.csv"
#> [204] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_177.csv"
#> [205] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_178.csv"
#> [206] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_179.csv"
#> [207] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_180.csv"
#> [208] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_181.csv"
#> [209] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_182.csv"
#> [210] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_183.csv"
#> [211] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_362.csv"
#> [212] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_363.csv"
#> [213] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_364.csv"
#> [214] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_365.csv"
#> [215] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_366.csv"
#> [216] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_367.csv"
#> [217] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_368.csv"
#> [218] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_369.csv"
#> [219] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_370.csv"
#> [220] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_371.csv"
#> [221] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_372.csv"
#> [222] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_373.csv"
#> [223] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_374.csv"
#> [224] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_375.csv"
#> [225] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_376.csv"
#> [226] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_377.csv"
#> [227] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_186.csv"
#> [228] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_187.csv"
#> [229] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_188.csv"
#> [230] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_189.csv"
#> [231] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_190.csv"
#> [232] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_191.csv"
#> [233] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_192.csv"
#> [234] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_193.csv"
#> [235] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_194.csv"
#> [236] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_195.csv"
#> [237] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_196.csv"
#> [238] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_197.csv"
#> [239] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_198.csv"
#> [240] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_199.csv"
#> [241] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_200.csv"
#> [242] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_201.csv"
#> [243] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_202.csv"
#> [244] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_203.csv"
#> [245] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_204.csv"
#> [246] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_205.csv"
#> [247] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_206.csv"
#> [248] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_207.csv"
#> [249] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_208.csv"
#> [250] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_209.csv"
#> [251] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_210.csv"
#> [252] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_211.csv"
#> [253] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_212.csv"
#> [254] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_213.csv"
#> [255] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_214.csv"
#> [256] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_215.csv"
#> [257] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_216.csv"
#> [258] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_217.csv"
#> [259] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_218.csv"
#> [260] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_219.csv"
#> [261] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_220.csv"
#> [262] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_378.csv"
#> [263] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_379.csv"
#> [264] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_380.csv"
#> [265] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_381.csv"
#> [266] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_382.csv"
#> [267] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_383.csv"
#> [268] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_384.csv"
#> [269] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_385.csv"
#> [270] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_386.csv"
#> [271] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_387.csv"
#> [272] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_388.csv"
#> [273] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_389.csv"
#> [274] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_390.csv"
#> [275] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_184.csv"
#> [276] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_185.csv"
#> [277] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_98.csv" 
#> [278] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_99.csv" 
#> [279] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_100.csv"
#> [280] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_101.csv"
#> [281] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_102.csv"
#> [282] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_103.csv"
#> [283] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_104.csv"
#> [284] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_105.csv"
#> [285] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_106.csv"
#> [286] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_107.csv"
#> [287] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_108.csv"
#> [288] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_109.csv"
#> [289] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_110.csv"
#> [290] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_111.csv"
#> [291] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_112.csv"
#> [292] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_113.csv"
#> [293] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_114.csv"
#> [294] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_92.csv" 
#> [295] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_93.csv" 
#> [296] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_94.csv" 
#> [297] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_95.csv" 
#> [298] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_96.csv" 
#> [299] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_97.csv" 
#> [300] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_74.csv" 
#> [301] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_75.csv" 
#> [302] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_76.csv" 
#> [303] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_77.csv" 
#> [304] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_78.csv" 
#> [305] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_79.csv" 
#> [306] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_80.csv" 
#> [307] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_81.csv" 
#> [308] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_82.csv" 
#> [309] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_83.csv" 
#> [310] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_1.csv"  
#> [311] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_2.csv"  
#> [312] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_3.csv"  
#> [313] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_4.csv"  
#> [314] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_5.csv"  
#> [315] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_6.csv"  
#> [316] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_7.csv"  
#> [317] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_8.csv"  
#> [318] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_9.csv"  
#> [319] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_10.csv" 
#> [320] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_11.csv" 
#> [321] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_12.csv" 
#> [322] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_13.csv" 
#> [323] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_14.csv" 
#> [324] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_15.csv" 
#> [325] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_16.csv" 
#> [326] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_17.csv" 
#> [327] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_18.csv" 
#> [328] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_19.csv" 
#> [329] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_20.csv" 
#> [330] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_21.csv" 
#> [331] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_22.csv" 
#> [332] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_23.csv" 
#> [333] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_24.csv" 
#> [334] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_25.csv" 
#> [335] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_26.csv" 
#> [336] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_27.csv" 
#> [337] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_28.csv" 
#> [338] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_29.csv" 
#> [339] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_30.csv" 
#> [340] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_31.csv" 
#> [341] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_32.csv" 
#> [342] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_33.csv" 
#> [343] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_34.csv" 
#> [344] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_35.csv" 
#> [345] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_36.csv" 
#> [346] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_37.csv" 
#> [347] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_38.csv" 
#> [348] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_39.csv" 
#> [349] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_40.csv" 
#> [350] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_58.csv" 
#> [351] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_59.csv" 
#> [352] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_60.csv" 
#> [353] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_61.csv" 
#> [354] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_62.csv" 
#> [355] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_63.csv" 
#> [356] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_64.csv" 
#> [357] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_65.csv" 
#> [358] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_66.csv" 
#> [359] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_67.csv" 
#> [360] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_68.csv" 
#> [361] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_69.csv" 
#> [362] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_70.csv" 
#> [363] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_71.csv" 
#> [364] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_72.csv" 
#> [365] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_73.csv" 
#> [366] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_84.csv" 
#> [367] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_85.csv" 
#> [368] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_86.csv" 
#> [369] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_87.csv" 
#> [370] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_88.csv" 
#> [371] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_89.csv" 
#> [372] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_90.csv" 
#> [373] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_91.csv" 
#> [374] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_391.csv"
#> [375] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_392.csv"
#> [376] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_393.csv"
#> [377] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_394.csv"
#> [378] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_395.csv"
#> [379] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_396.csv"
#> [380] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_41.csv" 
#> [381] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_42.csv" 
#> [382] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_43.csv" 
#> [383] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_44.csv" 
#> [384] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_45.csv" 
#> [385] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_46.csv" 
#> [386] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_47.csv" 
#> [387] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_48.csv" 
#> [388] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_49.csv" 
#> [389] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_50.csv" 
#> [390] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_51.csv" 
#> [391] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_52.csv" 
#> [392] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_53.csv" 
#> [393] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_54.csv" 
#> [394] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_55.csv" 
#> [395] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_56.csv" 
#> [396] "/tmp/Rtmp57MHSi/csv_dir_16c51848e157/trial_57.csv" 
#> 
#> Slot "template":
#>  [1] id                 trial_period       followup_time      outcome           
#>  [5] weight             treatment          catvarA            catvarB           
#>  [9] catvarC            nvarA              nvarB              nvarC             
#> [13] assigned_treatment
#> <0 rows> (or 0-length row.names)
#> 
#> Slot "N":
#> [1] 1939053
#> 

# delete after use
unlink(temp_dir, recursive = TRUE)