te_datastore_csv, functions and methods

This method is used internally by expand_trials to save the data to the "datastore" defined in set_expansion_options.

# S4 method for class 'te_datastore_csv'
save_expanded_data(object, data)

# S4 method for class 'te_datastore_csv'
read_expanded_data(object, period = NULL, subset_condition = NULL)

# S4 method for class 'te_datastore_csv'
sample_expanded_data(
  object,
  p_control,
  period = NULL,
  subset_condition = NULL,
  seed
)

Arguments

object

An object of class te_datastore or a child class.

data

A data frame containing the expanded trial data. The columns trial_period and id are present, which may be used in methods to save the data in an optimal way, such as with indexes, keys or separate files.

period

An integerish vector of non-zero length to select trial period(s) or NULL (default) to select all files.

subset_condition

A string of length 1 or NULL (default).

Value

A 'te_datastore_csv' object.

Slots

path

path to csv files.

files

names of csv files.

template

data.frame template.

Examples

temp_dir <- tempfile("csv_dir_")
dir.create(temp_dir)
datastore <- save_to_csv(temp_dir)
data(vignette_switch_data)
save_expanded_data(datastore, vignette_switch_data[1:200, ])
#> An object of class "te_datastore_csv"
#> Slot "path":
#> [1] "/tmp/RtmpXoZh1j/csv_dir_16fc724c65d0"
#> 
#> Slot "files":
#>                                                  file period
#> 1  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_261.csv    261
#> 2  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_262.csv    262
#> 3  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_263.csv    263
#> 4  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_264.csv    264
#> 5  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_265.csv    265
#> 6  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_266.csv    266
#> 7  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_267.csv    267
#> 8  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_268.csv    268
#> 9  /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_269.csv    269
#> 10 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_270.csv    270
#> 11 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_271.csv    271
#> 12 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_272.csv    272
#> 13 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_273.csv    273
#> 14 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_274.csv    274
#> 15 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_275.csv    275
#> 16 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_276.csv    276
#> 17 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_277.csv    277
#> 18 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_278.csv    278
#> 19 /tmp/RtmpXoZh1j/csv_dir_16fc724c65d0/trial_279.csv    279
#> 
#> Slot "template":
#>  [1] id                 trial_period       followup_time      outcome           
#>  [5] weight             treatment          catvarA            catvarB           
#>  [9] catvarC            nvarA              nvarB              nvarC             
#> [13] assigned_treatment
#> <0 rows> (or 0-length row.names)
#> 
#> Slot "N":
#> [1] 200
#> 

# delete after use
unlink(temp_dir, recursive = TRUE)