te_datastore_duckdb, functions and methods

This method is used internally by expand_trials to save the data to the "datastore" defined in set_expansion_options.

# S4 method for te_datastore_duckdb
save_expanded_data(object, data)

# S4 method for te_datastore_duckdb
read_expanded_data(object, period = NULL, subset_condition = NULL)

# S4 method for te_datastore_duckdb
sample_expanded_data(
  object,
  p_control,
  period = NULL,
  subset_condition = NULL,
  seed
)

Arguments

object

An object of class te_datastore or a child class.

data

A data frame containing the expanded trial data. The columns trial_period and id are present, which may be used in methods to save the data in an optimal way, such as with indexes, keys or separate files.

period

An integerish vector of non-zero length to select trial period(s) or NULL (default) to select all files.

subset_condition

A string of length 1 or NULL (default).

p_control

Probability of selecting a control.

seed

An integer seed or NULL (default).

Value

A 'te_datastore_duckdb' object.

Slots

path

Path to the duckdb file containing the data.

table

.

con

S4 object of class duckdb_connection.

Examples

temp_dir <- tempfile("csv_dir_")
dir.create(temp_dir)
datastore <- save_to_csv(temp_dir)
data(vignette_switch_data)
save_expanded_data(datastore, vignette_switch_data[1:200, ])
#> An object of class "te_datastore_csv"
#> Slot "path":
#> [1] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d"
#> 
#> Slot "files":
#>  [1] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_261.csv"
#>  [2] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_262.csv"
#>  [3] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_263.csv"
#>  [4] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_264.csv"
#>  [5] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_265.csv"
#>  [6] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_266.csv"
#>  [7] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_267.csv"
#>  [8] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_268.csv"
#>  [9] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_269.csv"
#> [10] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_270.csv"
#> [11] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_271.csv"
#> [12] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_272.csv"
#> [13] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_273.csv"
#> [14] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_274.csv"
#> [15] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_275.csv"
#> [16] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_276.csv"
#> [17] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_277.csv"
#> [18] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_278.csv"
#> [19] "/tmp/RtmpUSdUed/csv_dir_17155cf1d88d/trial_279.csv"
#> 
#> Slot "template":
#>  [1] id                 trial_period       followup_time      outcome           
#>  [5] weight             treatment          catvarA            catvarB           
#>  [9] catvarC            nvarA              nvarB              nvarC             
#> [13] assigned_treatment
#> <0 rows> (or 0-length row.names)
#> 
#> Slot "N":
#> [1] 200
#> 

# delete after use
unlink(temp_dir, recursive = TRUE)