Fit the marginal structural model for the sequence of emulated trials

Fit the marginal structural model for the sequence of emulated trials

fit_msm(
  object,
  use_sample_weights = TRUE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf)
)

# S4 method for class 'trial_sequence'
fit_msm(
  object,
  use_sample_weights = TRUE,
  analysis_weights = c("asis", "unweighted", "p99", "weight_limits"),
  weight_limits = c(0, Inf)
)

Arguments

object

A trial_sequence object

use_sample_weights

logical statement if sample weights should be used, default is TRUE

analysis_weights

a string, one of "asis"``, "unweighted", `"p99", "weight_limits".

weight_limits

a numeric vector of length 2 . If analysis_weights = "weight_limits this specifies the upper and lower limits for truncating the weights.

Before the outcome marginal structural model can be fit, the outcome model must be specified with set_outcome_model() and the data must be expanded into the trial sequence with expand_trials().

The model is fit based on the model_fitter specified in set_outcome_model using the internal fit_outcome_model method.

Value

A modified trial_sequence object with updated outcome_model slot.

Examples

trial_seq_object <- trial_sequence("ITT") |>
  set_data(data_censored) |>
  set_outcome_model(
    adjustment_terms = ~age_s,
    followup_time_terms = ~ stats::poly(followup_time, degree = 2)
  ) |>
  set_expansion_options(output = save_to_datatable(), chunk_size = 500) |>
  expand_trials() |>
  load_expanded_data()

fit_msm(trial_seq_object)
#> Trial Sequence Object 
#> Estimand: Intention-to-treat 
#> Data 
#> N: 725 observations from 89 patients 
#> Key: <id>
#>         id period treatment    x1           x2    x3        x4   age      age_s
#>      <int>  <int>     <num> <num>        <num> <int>     <num> <num>      <num>
#>   1:     1      0         1     1  1.146148362     0 0.7342030    36 0.08333333
#>   2:     1      1         1     1  0.002200337     0 0.7342030    37 0.16666667
#>  ---                                                                           
#> 724:    99      6         1     1 -0.033762356     1 0.5752681    71 3.00000000
#> 725:    99      7         0     0 -1.340496520     1 0.5752681    72 3.08333333
#>      outcome censored eligible time_of_event  first  am_1  cumA switch
#>        <num>    <int>    <num>         <num> <lgcl> <num> <num>  <num>
#>   1:       0        0        1          9999   TRUE     0     1      0
#>   2:       0        0        0          9999  FALSE     1     2      0
#>  ---                                                                  
#> 724:       0        0        0             7  FALSE     1     4      0
#> 725:       1        0        0             7  FALSE     1     4      1
#>      regime_start time_on_regime eligible0 eligible1
#>             <int>          <num>     <num>     <num>
#>   1:            0              0         1         0
#>   2:            0              1         0         1
#>  ---                                                
#> 724:            5              1         0         1
#> 725:            7              2         0         1
#>  
#> IPW for informative censoring: 
#>  - No weight model specified