Method to read, subset and sample expanded data

This method is used on trial_sequence objects to read, subset and sample expanded data.

sample_controls(
  object,
  p_control,
  period = NULL,
  subset_condition = NULL,
  seed = NULL
)

# S4 method for trial_sequence
sample_controls(
  object,
  p_control,
  period = NULL,
  subset_condition = NULL,
  seed = NULL
)

Arguments

object

An object of class trial_sequence.

p_control

Probability of selecting a control, default is 0.01.

period

An integerish vector of non-zero length to select trial period(s) or NULL (default) to select all trial periods.

subset_condition

A string or NULL (default). subset_condition will be translated to a call (in case the expanded data is saved as a data.table or in the csv format) or to a SQL-query (in case the expanded data is saved as a duckdb file).

The operators "==", "!=", ">", ">=", "<", "<=", %in%", "&", "|" are supported. Numeric vectors can be written as c(1, 2, 3) or 1:3. Variables are not supported.

Note: Make sure numeric vectors written as 1:3 are surrounded by spaces, e.g. a %in% c( 1:4 , 6:9 ), otherwise the code will fail.

seed

An integer seed or NULL (default).

Note: The same seed will return a different result depending on the class of the te_datastore object contained in the trial_sequence object.

Value

A data.frame of class data.table.

Examples

# create a trial_sequence-class object
trial_itt_dir <- file.path(tempdir(), "trial_itt")
dir.create(trial_itt_dir)
trial_itt <- trial_sequence(estimand = "ITT") |>
  set_data(data = data_censored) |>
  set_outcome_model(adjustment_terms = ~ x1 + x2)

trial_itt_csv <- set_expansion_options(
  trial_itt,
  output = save_to_csv(file.path(trial_itt_dir, "trial_csvs")),
  chunk_size = 500
) |>
  expand_trials()

# sample_controls default behaviour returns all trial_periods
sample_controls(trial_itt_csv, p_control = 0.01)
#>        id trial_period followup_time outcome weight treatment    x1          x2
#>     <int>        <int>         <int>   <int>  <int>     <int> <int>       <num>
#>  1:    15            0             0       1      1         1     0 -0.73652563
#>  2:    32            0             0       1      1         1     1  1.98613804
#>  3:    23            0             0       0      1         1     0 -0.61336220
#>  4:    35            1             0       0     NA         0     0 -0.29198761
#>  5:    54            6             0       0     NA         0     0 -0.07989556
#>  6:    54           17             0       0     NA         1     0  1.84642254
#>  7:    63            0             1       0     NA         1     0 -0.14360058
#>  8:    57            1             1       0     NA         0     1  0.13248495
#>  9:    74            3             1       0     NA         0     0 -0.84801944
#> 10:    61            0             2       1     NA         0     1 -0.61233242
#> 11:    68            0             4       1     NA         0     0 -0.39198375
#> 12:    54           15             4       0     NA         0     0  0.68507216
#> 13:    29            0             5       0     NA         0     0  0.40372402
#> 14:    99            0             6       0     NA         1     1 -0.34637784
#> 15:    49            1             6       1     NA         0     0 -2.26870654
#> 16:    29            0             7       1     NA         1     0  0.40372402
#> 17:    49            0             7       1     NA         0     0  1.15721110
#> 18:    89            0             7       1     NA         0     1  0.41285176
#> 19:    99            0             7       1     NA         0     1 -0.34637784
#> 20:    38            1             7       1     NA         0     1 -1.39123463
#> 21:    54           11             7       0     NA         0     1 -1.01922545
#> 22:    38            0             8       1     NA         0     0 -0.24691109
#> 23:    21            3             8       0     NA         0     0  1.87509512
#> 24:    71            1            13       1     NA         0     0 -0.39700243
#> 25:    71            0            14       1     NA         0     0 -0.26789128
#> 26:    98            0            14       1     NA         0     1  1.39233946
#> 27:    55            0            14       0     NA         0     1  0.44800965
#> 28:    59            0            15       0     NA         1     0 -0.53973276
#> 29:    44            1            15       0     NA         0     0 -0.60902320
#>        id trial_period followup_time outcome weight treatment    x1          x2
#>     assigned_treatment sample_weight
#>                  <int>         <num>
#>  1:                  1             1
#>  2:                  1             1
#>  3:                  1           100
#>  4:                  0           100
#>  5:                  0           100
#>  6:                  1           100
#>  7:                  1           100
#>  8:                  0           100
#>  9:                  0           100
#> 10:                  1             1
#> 11:                  1             1
#> 12:                  0           100
#> 13:                  1           100
#> 14:                  1           100
#> 15:                  1             1
#> 16:                  1             1
#> 17:                  0             1
#> 18:                  1             1
#> 19:                  1             1
#> 20:                  1             1
#> 21:                  0           100
#> 22:                  0             1
#> 23:                  1           100
#> 24:                  1             1
#> 25:                  0             1
#> 26:                  1             1
#> 27:                  1           100
#> 28:                  0           100
#> 29:                  0           100
#>     assigned_treatment sample_weight

# sample_controls can subset the data before sampling
sample_controls(
  trial_itt_csv,
  p_control = 0.2,
  period = 1:20,
  subset_condition = "followup_time %in% 1:20 & treatment == 1",
)
#>        id trial_period followup_time outcome weight treatment    x1          x2
#>     <int>        <int>         <int>   <int> <lgcl>     <int> <int>       <num>
#>  1:    20            1             1       0     NA         1     0 -0.42516753
#>  2:    59            1             1       0     NA         1     0 -1.18282317
#>  3:    60            1             1       0     NA         1     0  0.54368336
#>  4:    95            2             1       0     NA         1     1  0.01992213
#>  5:    16            1             2       0     NA         1     1 -0.69264361
#>  6:    95            1             2       0     NA         1     0 -1.52782784
#>  7:    60            2             2       0     NA         1     1  0.79983377
#>  8:    44            4             2       0     NA         1     0 -0.41225502
#>  9:    38            1             3       0     NA         1     1 -1.39123463
#> 10:    34            1             3       0     NA         1     1  0.24167648
#> 11:    14            1             3       0     NA         1     0  0.51664948
#> 12:    16            1             3       0     NA         1     1 -0.69264361
#> 13:    44            1             3       0     NA         1     0 -0.60902320
#> 14:    53            1             3       0     NA         1     1 -0.18557643
#> 15:    54           14             3       0     NA         1     1  0.54606151
#> 16:    34            1             4       0     NA         1     1  0.24167648
#> 17:    71            1             4       0     NA         1     0 -0.39700243
#> 18:    27            2             4       0     NA         1     0  1.75539753
#> 19:    54           13             4       0     NA         1     0 -2.42496589
#> 20:    60            1             5       0     NA         1     0  0.54368336
#> 21:    27            1             5       0     NA         1     0  0.68971049
#> 22:    74            4             5       0     NA         1     1 -0.97471771
#> 23:    44            4             5       0     NA         1     0 -0.41225502
#> 24:    95            1             6       0     NA         1     0 -1.52782784
#> 25:    33            1             6       0     NA         1     0 -0.55141438
#> 26:    59            1             6       0     NA         1     0 -1.18282317
#> 27:    74            2             6       0     NA         1     0  0.11750682
#> 28:    33            1             7       0     NA         1     0 -0.55141438
#> 29:    71            1             7       0     NA         1     0 -0.39700243
#> 30:    44            1             7       0     NA         1     0 -0.60902320
#> 31:    44            2             7       0     NA         1     0  0.84818522
#> 32:    59            1             8       0     NA         1     0 -1.18282317
#> 33:    53            1             9       0     NA         1     1 -0.18557643
#> 34:    95            2            10       0     NA         1     1  0.01992213
#> 35:    44            3            10       0     NA         1     1 -0.59779120
#> 36:    44            2            11       0     NA         1     0  0.84818522
#> 37:    71            1            12       0     NA         1     0 -0.39700243
#> 38:    53            1            12       0     NA         1     1 -0.18557643
#> 39:    95            2            12       0     NA         1     1  0.01992213
#> 40:    95            3            12       0     NA         1     1  0.68712558
#> 41:    44            3            12       0     NA         1     1 -0.59779120
#> 42:    44            2            13       0     NA         1     0  0.84818522
#> 43:    44            1            14       0     NA         1     0 -0.60902320
#> 44:    14            1            14       0     NA         1     0  0.51664948
#> 45:    33            1            15       0     NA         1     0 -0.55141438
#> 46:    95            1            15       0     NA         1     0 -1.52782784
#> 47:    96            1            15       0     NA         1     0 -1.08532471
#> 48:    95            1            16       0     NA         1     0 -1.52782784
#> 49:    59            2            16       0     NA         1     1  0.68201510
#> 50:    96            1            17       0     NA         1     0 -1.08532471
#> 51:    59            1            17       0     NA         1     0 -1.18282317
#>        id trial_period followup_time outcome weight treatment    x1          x2
#>     assigned_treatment sample_weight
#>                  <int>         <num>
#>  1:                  0             5
#>  2:                  0             5
#>  3:                  0             5
#>  4:                  0             5
#>  5:                  1             5
#>  6:                  0             5
#>  7:                  1             5
#>  8:                  1             5
#>  9:                  1             5
#> 10:                  1             5
#> 11:                  1             5
#> 12:                  1             5
#> 13:                  0             5
#> 14:                  1             5
#> 15:                  0             5
#> 16:                  1             5
#> 17:                  1             5
#> 18:                  0             5
#> 19:                  0             5
#> 20:                  0             5
#> 21:                  0             5
#> 22:                  0             5
#> 23:                  1             5
#> 24:                  0             5
#> 25:                  1             5
#> 26:                  0             5
#> 27:                  0             5
#> 28:                  1             5
#> 29:                  1             5
#> 30:                  0             5
#> 31:                  0             5
#> 32:                  0             5
#> 33:                  1             5
#> 34:                  0             5
#> 35:                  0             5
#> 36:                  0             5
#> 37:                  1             5
#> 38:                  1             5
#> 39:                  0             5
#> 40:                  1             5
#> 41:                  0             5
#> 42:                  0             5
#> 43:                  0             5
#> 44:                  1             5
#> 45:                  1             5
#> 46:                  0             5
#> 47:                  1             5
#> 48:                  0             5
#> 49:                  1             5
#> 50:                  1             5
#> 51:                  0             5
#>     assigned_treatment sample_weight

# delete after use
unlink(trial_itt_dir, recursive = TRUE)